python - move thrift testing util to reusable location
